+++ tests/test_tornado_client.py
@@ -1,15 +1,15 @@
 import tornado.gen
 import tornado.ioloop
+import vcr_unittest
 
 from deezer import Album
 from deezer.contrib.tornado import AsyncClient
-from .base import BaseTestCase
 
 
-class TestAsyncClient(BaseTestCase):
+class TestAsyncClient(vcr_unittest.VCRTestCase):
     def setUp(self):
         super().setUp()
-        self.client = AsyncClient()
+        self.client = AsyncClient(do_not_compress_reponse=False)
 
     def test_get_object(self):
         @tornado.gen.coroutine